requirements.txt
pillow
django-rest-framework